feat: UserProfile 클래스 추출 및 초기화 로직 캡슐화
